requirements.txt
numpy
pandas
yfinance
streamlit
matplotlib
google-cloud-firestore
google-auth